move idx parsing to vob module
