+++ src/mkv.rs
@@ -9,7 +9,11 @@ use webm_iterable::{
     WebmIterator,
 };
 
-use crate::{pgs, text::sanitize_text, vob};
+use crate::{
+    pgs,
+    text::sanitize_text,
+    vob::{self, parse_idx},
+};
 
 #[derive(Debug, PartialEq, Clone)]
 pub enum KnownLanguage {
@@ -57,72 +61,7 @@ impl KnownEncoding {
             "S_HDMV/PGS" => KnownEncoding::PGS,
             "S_VOBSUB" => {
                 if let Some(data) = data {
-                    let idx_string = String::from_utf8_lossy(data);
-                    //println!("{}", idx_string);
-                    let lines = idx_string.lines();
-                    //let first_line = lines.nth(0).unwrap();
-                    //if first_line != r#"# VobSub index file, v7 (do not modify this line!)"# {
-                    //    println!("Warning! Expected to see the VobSub v7 line at the beginning of the private data...");
-                    //}
-                    let mut size = None;
-                    let mut palette = None;
-                    for line in lines {
-                        // Skip comments
-                        if line.starts_with("#") {
-                            continue;
-                        }
-
-                        // Split the line on the first ':'
-                        if let Some((name, value)) = line.split_once(':') {
-                            let value = value.trim();
-                            match name {
-                                "size" => {
-                                    let (width_str, height_str) = value.split_once('x').unwrap();
-                                    let width = u32::from_str_radix(width_str, 10).unwrap();
-                                    let height = u32::from_str_radix(height_str, 10).unwrap();
-                                    size = Some((width, height));
-                                }
-                                "palette" => {
-                                    let mut colors = Vec::new();
-                                    let color_strs = value.split(", ");
-                                    for color_str in color_strs {
-                                        assert_eq!(color_str.len(), 6);
-                                        // Not sure what the format is, assuming RGB for now
-                                        let r_str = &color_str[0..2];
-                                        let g_str = &color_str[2..4];
-                                        let b_str = &color_str[4..6];
-
-                                        let r = u8::from_str_radix(r_str, 16).unwrap();
-                                        let g = u8::from_str_radix(g_str, 16).unwrap();
-                                        let b = u8::from_str_radix(b_str, 16).unwrap();
-
-                                        let color = Color {
-                                            A: 255,
-                                            R: r,
-                                            G: g,
-                                            B: b,
-                                        };
-                                        colors.push(color);
-                                    }
-                                    palette = Some(colors);
-                                }
-                                _ => {
-                                    //println!("Unknown name: \"{}\"", name);
-                                }
-                            }
-                        }
-                    }
-
-                    let (width, height) =
-                        size.expect("Expected size in Vob subtitle track private data");
-                    let palette =
-                        palette.expect("Expected palette in Vob subtitle track private data");
-
-                    KnownEncoding::VOB {
-                        width,
-                        height,
-                        palette,
-                    }
+                    parse_idx(data)
                 } else {
                     panic!("Expected private data for VOB subtitles!");
                 }
